Refactor to add support for faux_detector with Yolov5
